bazaar_upgraded/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1

python-dotenv==1.0.1
SQLAlchemy==2.0.36
cryptography==43.0.3
authlib>=1.2.0
Pillow>=10.0.0
requests>=2.31.0
PyMySQL==1.1.1
gunicorn
